tambahan bagian admin

+++ admin/admincineverse/admin.php
@@ -1,10 +1,20 @@
+<?php
+session_start();
+
+// Check if the user is logged in
+if (!isset($_SESSION["username"])) {
+    // Redirect to the login page or perform other actions
+    header("Location: login.php");
+    exit();
+}
+?>
 <!DOCTYPE html>
 <html lang="en">
 
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>Admin</title>
+    <title>Cineverse | Admin</title>
     <meta content="" name="description">
     <meta content="" name="keywords">
 
@@ -14,9 +24,7 @@
 
     <!-- Google Fonts -->
     <link href="https://fonts.gstatic.com" rel="preconnect">
-    <link
-        href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i|Nunito:300,300i,400,400i,600,600i,700,700i|Poppins:300,300i,400,400i,500,500i,600,600i,700,700i"
-        rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Open+Sans:300,300i,400,400i,600,600i,700,700i|Nunito:300,300i,400,400i,600,600i,700,700i|Poppins:300,300i,400,400i,500,500i,600,600i,700,700i" rel="stylesheet">
 
     <!-- Vendor CSS Files -->
     <link href="../assets/vendor/bootstrap/css/bootstrap.min.css" rel="stylesheet">
@@ -33,14 +41,6 @@
 
     <!-- Template Main CSS File -->
     <link href="../assets/css/style.css" rel="stylesheet">
-    <!-- =======================================================
-  * Template Name: NiceAdmin
-  * Updated: Nov 17 2023 with Bootstrap v5.3.2
-  * Template URL: https://bootstrapmade.com/nice-admin-bootstrap-admin-html-template/
-  * Author: BootstrapMade.com
-  * License: https://bootstrapmade.com/license/
-  ======================================================== -->
-
 </head>
 
 <body>
@@ -62,44 +62,19 @@
                 <i class="bi bi-list toggle-sidebar-btn"></i>
             </div><!-- End Logo -->
 
-            <div class="search-bar">
-                <form class="search-form d-flex align-items-center" method="POST" action="#">
-                    <input type="text" name="query" placeholder="Search" title="Enter search keyword">
-                    <button type="submit" title="Search"><i class="bi bi-search"></i></button>
-                </form>
-            </div><!-- End Search Bar -->
-
             <nav class="header-nav ms-auto">
                 <ul class="d-flex align-items-center">
-
-                    <li class="nav-item d-block d-lg-none">
-                        <a class="nav-link nav-icon search-bar-toggle " href="#">
-                            <i class="bi bi-search"></i>
-                        </a>
-                    </li><!-- End Search Icon-->
-
                     <li class="nav-item dropdown pe-3">
-
-                        <a class="nav-link nav-profile d-flex align-items-center pe-0" href="#"
-                            data-bs-toggle="dropdown">
-                            <img src="../assets/img/profile-img.jpg" alt="Profile" class="rounded-circle">
-                            <span class="d-none d-md-block dropdown-toggle ps-2">K. Anderson</span>
+                        <a class="nav-link nav-profile d-flex align-items-center pe-0" href="#" data-bs-toggle="dropdown">
+                            <span class="d-none d-md-block dropdown-toggle ps-2"><?php echo $_SESSION["username"] ?></span>
                         </a><!-- End Profile Iamge Icon -->
 
                         <ul class="dropdown-menu dropdown-menu-end dropdown-menu-arrow profile">
                             <li class="dropdown-header">
-                                <h6>Kevin Anderson</h6>
-                                <span>Web Designer</span>
+                                <h6><?php echo $_SESSION["username"] ?></h6>
                             </li>
                             <li>
                                 <hr class="dropdown-divider">
-                            </li>
-
-                            <li>
-                                <a class="dropdown-item d-flex align-items-center" href="users-profile.html">
-                                    <i class="bi bi-person"></i>
-                                    <span>My Profile</span>
-                                </a>
                             </li>
                             <li>
                                 <a class="dropdown-item d-flex align-items-center" href="login.php">
@@ -117,63 +92,61 @@
         </header><!-- End Header -->
 
         <!-- ======= Sidebar ======= -->
-    <aside id="sidebar" class="sidebar">
-        <ul class="sidebar-nav" id="sidebar-nav">
-            <li class="nav-item">
-                <a class="nav-link " href="dashboard.php">
-                    <i class="bi bi-grid"></i>
-                    <span>Dashboard</span>
-                </a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link collapsed" data-bs-target="#components-nav" data-bs-toggle="collapse"
-                    href="dashboard.php">
-                    <i class="bi bi-menu-button-wide"></i><span>Film</span><i class="bi bi-chevron-down ms-auto"></i>
-                </a>
-                <ul id="components-nav" class="nav-content collapse " data-bs-parent="#sidebar-nav">
-                    <li>
-                        <a href="film/film.php">
-                            <i class="bi bi-circle"></i><span>Data film</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="genre/genre.php">
-                            <i class="bi bi-circle"></i><span>Genre Film</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="pemain/pemain.php">
-                            <i class="bi bi-circle"></i><span>Daftar Pemain</span>
-                        </a>
-                    </li>
-                </ul>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link collapsed" data-bs-target="#tables-nav" data-bs-toggle="collapse"
-                    href="dashboard.php">
-                    <i class="bi bi-menu-button-wide"></i><span>User</span><i class="bi bi-chevron-down ms-auto"></i>
-                </a>
-                <ul id="tables-nav" class="nav-content collapse " data-bs-parent="#sidebar-nav">
-                    <li>
-                        <a href="user/user.php">
-                            <i class="bi bi-circle"></i><span>User Cineverse</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="komentar/komentar.php">
-                            <i class="bi bi-circle"></i><span>Review</span>
-                        </a>
-                    </li>
-                </ul>
-            </li><!-- End Forms Nav -->
-            <li class="nav-item">
-                <a class="nav-link" href="admincineverse/admin.php">
-                    <i class="bi bi-grid"></i>
-                    <span>Admin Cineverse</span>
-                </a>
-            </li>
-        </ul>
-    </aside><!-- End Sidebar-->
+        <aside id="sidebar" class="sidebar">
+            <ul class="sidebar-nav" id="sidebar-nav">
+                <li class="nav-item">
+                    <a class="nav-link " href="../dashboard.php">
+                        <i class="bi bi-grid"></i>
+                        <span>Dashboard</span>
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link collapsed" data-bs-target="#components-nav" data-bs-toggle="collapse" href="dashboard.php">
+                        <i class="bi bi-menu-button-wide"></i><span>Film</span><i class="bi bi-chevron-down ms-auto"></i>
+                    </a>
+                    <ul id="components-nav" class="nav-content collapse " data-bs-parent="#sidebar-nav">
+                        <li>
+                            <a href="../film/film.php">
+                                <i class="bi bi-circle"></i><span>Data film</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="../genre/genre.php">
+                                <i class="bi bi-circle"></i><span>Genre Film</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="../pemain/pemain.php">
+                                <i class="bi bi-circle"></i><span>Daftar Pemain</span>
+                            </a>
+                        </li>
+                    </ul>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link collapsed" data-bs-target="#tables-nav" data-bs-toggle="collapse" href="dashboard.php">
+                        <i class="bi bi-menu-button-wide"></i><span>User</span><i class="bi bi-chevron-down ms-auto"></i>
+                    </a>
+                    <ul id="tables-nav" class="nav-content collapse " data-bs-parent="#sidebar-nav">
+                        <li>
+                            <a href="../user/user.php">
+                                <i class="bi bi-circle"></i><span>User Cineverse</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="../komentar/komentar.php">
+                                <i class="bi bi-circle"></i><span>Review</span>
+                            </a>
+                        </li>
+                    </ul>
+                </li><!-- End Forms Nav -->
+                <li class="nav-item">
+                    <a class="nav-link" href="admincineverse/admin.php">
+                        <i class="bi bi-grid"></i>
+                        <span>Admin Cineverse</span>
+                    </a>
+                </li>
+            </ul>
+        </aside><!-- End Sidebar-->
 
         <main id="main" class="main">
 
@@ -182,7 +155,7 @@
                 <nav>
                     <ol class="breadcrumb">
                         <li class="breadcrumb-item"><a href="../dashboard.php">Dashboard</a></li>
-                        <li class="breadcrumb-item active"><a href="genre.php">Data Admin</a></li>
+                        <li class="breadcrumb-item active">Admin Cineverse</li>
                     </ol>
                 </nav>
             </div><!-- End Page Title -->
@@ -190,7 +163,7 @@
 
             <!-- Table -->
             <div class="table-responsive mt-4">
-                <table class="table table-striped table-bordered" id="film" class="display" style="width:100%">
+                <table class="table table-striped table-bordered" id="data-tabel" class="display" style="width:100%">
                     <a href="tambah.php?" class="btn btn-light" style="margin-bottom:5px"><i class="fa-solid fa-user-plus"></i></a>
                     <thead>
                         <tr>
@@ -202,28 +175,32 @@
                     </thead>
 
                     <tbody> <?php
-                        if (mysqli_num_rows($query) > 0) {
-                            while ($data = mysqli_fetch_array($query)) {
-                        ?> <tr>
-                            <td> <?php echo $data["id"] ?></td>
-                            <td> <?php echo $data["username"] ?></td>
-                            <td> <?php echo $data["password"] ?></td>
-                            <td> <a class="btn btn-warning" href="edit.php?id=<?php echo $data["id"] ?>">
-                                    <i class="fa-solid fa-pen-to-square"></i></a>
-                                <a class="btn btn-danger" href="proses_hapus.php?id=<?php echo $data["id"] ?>"
-                                    onclick="return confirm('Yakin Data Akan Dihapus?')">
-                                    <i class="fa-solid fa-trash-can"></i> </a>
-                            </td>
-                        </tr>
-                        <?php
-                            } ?>
+                            if (mysqli_num_rows($query) > 0) {
+                                $no = 1;
+                                while ($data = mysqli_fetch_array($query)) {
+                            ?> <tr>
+                                    <td> <?php echo $no ?></td>
+                                    <td> <?php echo $data["username"] ?></td>
+                                    <td> <?php
+                                            $password = $data['password'];
+                                            $length = strlen($password);
+
+                                            // Ganti setiap karakter dengan "*"
+                                            $maskedPassword = str_repeat('*', $length);
+
+                                            // Tampilkan password yang telah dimask
+                                            echo $maskedPassword;
+                                            ?>
+                                    </td>
+                                    <td> <a class="btn btn-warning" href="edit.php?id=<?php echo $data["id"] ?>">
+                                            <i class="fa-solid fa-pen-to-square"></i></a>
+                                        <a class="btn btn-danger" href="proses_hapus.php?id=<?php echo $data["id"] ?>" onclick="return confirm('Yakin Data Akan Dihapus?')">
+                                            <i class="fa-solid fa-trash-can"></i> </a>
+                                    </td>
+                                </tr>
+                            <?php $no++;
+                                } ?>
                         <?php } ?>
-                        <!-- Datatables -->
-                        <script>
-                        $(document).ready(function() {
-                            $('#film').DataTable();
-                        });
-                        </script>
                     </tbody>
                 </table>
             </div>
@@ -235,19 +212,18 @@
     <!-- ======= Footer ======= -->
     <footer id="footer" class="footer fixed-bottom">
         <div class="copyright">
-            &copy; Copyright <strong><span>NiceAdmin</span></strong>. All Rights Reserved
+            &copy; Copyright <strong><span>Cineverse Admin</span></strong>. All Rights Reserved
         </div>
         <div class="credits">
             <!-- All the links in the footer should remain intact. -->
             <!-- You can delete the links only if you purchased the pro version. -->
             <!-- Licensing information: https://bootstrapmade.com/license/ -->
             <!-- Purchase the pro version with working PHP/AJAX contact form: https://bootstrapmade.com/nice-admin-bootstrap-admin-html-template/ -->
-            Designed by <a href="https://bootstrapmade.com/">BootstrapMade</a>
+            Designed by <strong><span>Cineverse</span></strong>
         </div>
     </footer><!-- End Footer -->
 
-    <a href="#" class="back-to-top d-flex align-items-center justify-content-center"><i
-            class="bi bi-arrow-up-short"></i></a>
+    <a href="#" class="back-to-top d-flex align-items-center justify-content-center"><i class="bi bi-arrow-up-short"></i></a>
 
     <!-- Vendor JS Files -->
     <script src="../assets/vendor/apexcharts/apexcharts.min.js"></script>
@@ -266,25 +242,25 @@
     <script src="https://cdn.datatables.net/1.13.7/js/jquery.dataTables.min.js"></script>
     <script src="https://cdn.datatables.net/1.13.7/js/dataTables.bootstrap5.min.js"></script>
     <script>
-    new DataTable('#data-tabel');
+        new DataTable('#data-tabel');
     </script>
     <script>
-    $(document).ready(function() {
-        // Menangani klik tombol "Hapus"
-        $('.btn-delete').on('click', function(e) {
-            e.preventDefault(); // Mencegah tindakan asli tautan
+        $(document).ready(function() {
+            // Menangani klik tombol "Hapus"
+            $('.btn-delete').on('click', function(e) {
+                e.preventDefault(); // Mencegah tindakan asli tautan
 
-            var id = $(this).data('id');
-            var confirmation = confirm("Apakah Anda yakin ingin menghapus produk ini?");
+                var id = $(this).data('id');
+                var confirmation = confirm("Apakah Anda yakin ingin menghapus produk ini?");
 
-            if (confirmation) {
-                // Jika pengguna mengonfirmasi, arahkan ke halaman proses penghapusan
-                window.location.href = "proses_hapus.php?id=" + id;
-            } else {
-                // Jika pengguna membatalkan, tidak terjadi apa-apa
-            }
+                if (confirmation) {
+                    // Jika pengguna mengonfirmasi, arahkan ke halaman proses penghapusan
+                    window.location.href = "proses_hapus.php?id=" + id;
+                } else {
+                    // Jika pengguna membatalkan, tidak terjadi apa-apa
+                }
+            });
         });
-    });
     </script>
 
 </body>
